requirements.txt
# requirements.in  (top-level only)
fastapi>=0.110,<1
uvicorn>=0.29,<1
pydantic>=2,<3        # FastAPI now targets Pydantic v2
sentence-transformers>=2.5,<3
faiss-cpu>=1.7,<2
prometheus-client>=0.17,<1
openai>=1,<2          # optional; safe modern API
pypdf>=4,<5           # local PDF text extraction (no OCR)
pymupdf>=1.24,<2      # local PDF rendering for OCR fallback
pytesseract>=0.3,<1   # OCR interface (requires system tesseract)
Pillow>=10,<11        # image handling for OCR
pandas>=2,<3          # parquet IO for probe/cluster/route artifacts
pyarrow>=16,<17       # parquet engine
scikit-learn>=1.4,<2  # clustering and scaling
opencv-python-headless>=4.9,<5  # image metrics (laplacian)
